requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
azure-core==1.26.1
azure-storage-blob==12.14.1
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
cssselect==1.2.0
Django==4.1.4
django-appconf==1.0.5
django-celery-email==3.0.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-newsletter==0.9.1
django-rq==2.4.1
django-storages==1.13.1
django-tinymce==3.5.0
django-translation-manager==1.3.0
djangorestframework==3.14.0
Faker==15.3.4
gunicorn==20.1.0
idna==3.4
isodate==0.6.1
kombu==5.2.4
ldif3==3.1.1
lxml==4.9.1
Markdown==3.4.1
msrest==0.7.1
oauthlib==3.2.2
Pillow==9.3.0
polib==1.1.1
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycparser==2.21
Pygments==2.13.0
pyquery==1.4.3
python-card-me==0.9.3
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rq==1.11.1
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
surlex==0.2.0
typing_extensions==4.4.0
unicodecsv==0.14.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
